Removed (I)DbTable class and now inheriting from Model<> in database table models.
