+++ IdeallyConnected.Database/TestDatabases/SampleSuperstore/Return.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 
 namespace IdeallyConnected.TestDatabases
 {
-    public class Return : DbTable
+    public class Return : Model<Return>
     {
         public string OrderID { get; set; }
         public bool Returned { get; set; }
